[FEAT] 팔로우, 팔로우 취소 전 팔로우 유무 확인

+++ graphQL/src/main/kotlin/com/sopt/team1/service/FollowService.kt
@@ -12,36 +12,52 @@ import org.springframework.stereotype.Service
 class FollowService(private val followRepository: FollowRepository, private val memberRepository: MemberRepository) {
     @Transactional
     fun follow(memberId: Long, targetId: Long): FollowResponseDTO {
-        val follow = Follow(followerId = memberId, followingId = targetId)
-        followRepository.save(follow)
+        when {
+            memberId == targetId -> throw IllegalArgumentException("You can't follow yourself")
+            followRepository.findByFollowerIdAndFollowingId(memberId, targetId) -> throw IllegalArgumentException("You already follow this member")
+            else -> {
+                val follow = Follow(followerId = memberId, followingId = targetId)
+                followRepository.save(follow)
 
-        val follower = memberRepository.findById(memberId).orElseThrow { throw EntityNotFoundException("Member not found with id: $memberId") }
-        follower.following += 1
+                val follower = memberRepository.findById(memberId)
+                    .orElseThrow { throw EntityNotFoundException("Member not found with id: $memberId") }
+                follower.following += 1
 
-        val target = memberRepository.findById(targetId).orElseThrow { throw EntityNotFoundException("Member not found with id: $targetId") }
-        target.follower += 1
+                val target = memberRepository.findById(targetId)
+                    .orElseThrow { throw EntityNotFoundException("Member not found with id: $targetId") }
+                target.follower += 1
 
-        return FollowResponseDTO(
-            id = follower.id ?: error("Follower id should not be null"),
-            followerCount = follower.follower,
-            followingCount = follower.following
-        )
+                return FollowResponseDTO(
+                    id = follower.id ?: error("Follower id should not be null"),
+                    followerCount = follower.follower,
+                    followingCount = follower.following
+                )
+            }
+        }
     }
 
     @Transactional
     fun unfollow(memberId: Long, targetId: Long): FollowResponseDTO{
-        followRepository.deleteByFollowerIdAndFollowingId(memberId, targetId)
+        when{
+            memberId == targetId -> throw IllegalArgumentException("You can't unfollow yourself")
+            !followRepository.findByFollowerIdAndFollowingId(memberId, targetId) -> throw IllegalArgumentException("You don't follow this member")
+            else -> {
+                followRepository.deleteByFollowerIdAndFollowingId(memberId, targetId)
 
-        val follower = memberRepository.findById(memberId).orElseThrow { throw EntityNotFoundException("Member not found with id: $memberId") }
-        follower.following -= 1
+                val follower = memberRepository.findById(memberId)
+                    .orElseThrow { throw EntityNotFoundException("Member not found with id: $memberId") }
+                follower.following -= 1
 
-        val target = memberRepository.findById(targetId).orElseThrow { throw EntityNotFoundException("Member not found with id: $targetId") }
-        target.follower -= 1
+                val target = memberRepository.findById(targetId)
+                    .orElseThrow { throw EntityNotFoundException("Member not found with id: $targetId") }
+                target.follower -= 1
 
-        return FollowResponseDTO(
-            id = follower.id ?: error("Follower id should not be null"),
-            followerCount = follower.follower,
-            followingCount = follower.following
-        )
+                return FollowResponseDTO(
+                    id = follower.id ?: error("Follower id should not be null"),
+                    followerCount = follower.follower,
+                    followingCount = follower.following
+                )
+            }
+        }
     }
 }
